Refactor ParaManager
